requirements.txt
setuptools<70
tensorflow
tensorflow-hub
gymnasium
playwright
jinja2
numpy
Pillow
pytest
pytest-asyncio
ruff
python-dotenv
